refactor: package 명 user -> member 변경
